Fixed: Don't fail if items contain \r.
First, as I write some of the items on Windows (and sometimes I'm in a
hurry), I can't guarantee that all of my files will use Unix line
endings. So, we need to remove those nasty \r's.

Second, the first loop read the file and the second loop read it again
(in order to show the subject line). That's not nice -- additionally,
the stripping of \r's would have had to happen *twice*. Now, the file is
only read once.

Third, "sed q" is a lot shorter than "sed -n '1p;q'".