Pick random IDs and use complete_id more wisely.
That's funny. I'm using this little tool for quite some time now, but
I've never stumbled upon this rather obvious bug.

Imagine you add a new item on computer A and (in parallel) a new item on
computer B. The old code assigned the same ID to both these items. Sure,
that's a conflict. And you have to resolve it manually.

Now, we pick the IDs randomly. Still it's not 100% sure that you do NOT
get a conflict -- but it's much less probable.