Introduced GITODO_DATA, use XDG_* or HOME otherwise.
Historically, gitodo was meant to have the items repository in a
subdirectory of the code repo. This was a special requirement because I
had to use Windows and this simplified things a lot. This is no longer
relevant for me.

The new mechanism is cleaner, more flexible and more compliant to
standards.