Fixed: Wrong usage of "next".
Misread the docs. I even wondered about it when I wrote it because
"next" to *end* a block didn't make any sense. But hey, strange things
happen from time to time...

There's no change in behaviour, though. "next" moves to the next filter
but "match" has already been set to "true". We didn't get any false
results with the old code but it was wrong nevertheless.