Remove comment about "global todo items"
That comment has been there since forever. It won't change any time
soon. I also think that the code is okay-ish enough at the moment. We
don't need Java-like comment bloat everywhere. What we need is comments
that explain the intentions of code *if* the code doesn't speak for
itself.